feat: Add Team and User Management Pages
- Implemented TeamViewDelete component for deleting teams.
- Created TeamViewIndex component to display team details.
- Added TeamViewLayout for team management layout.
- Developed TeamViewMembers component to manage team members.
- Introduced TeamViewPermissions for managing team permissions.
- Created TeamViewSideMenu for navigation within team settings.
- Implemented UserViewCustomFields for displaying user custom fields.
- Added UserViewDelete component for removing users from projects.
- Created UserViewIndex to show user details.
- Developed UserViewLayout for user management layout.
- Implemented UserViewSideMenu for navigation within user settings.
- Added UserViewTeams component to manage user team memberships.
