+++ App/FeatureSet/Dashboard/src/Utils/Breadcrumbs/SettingsBreadcrumbs.ts
@@ -34,6 +34,38 @@ export function getSettingsBreadcrumbs(path: string): Array<Link> | undefined {
       "Teams",
       "View Team",
     ]),
+    ...BuildBreadcrumbLinksByTitles(PageMap.SETTINGS_TEAM_VIEW_MEMBERS, [
+      "Project",
+      "Settings",
+      "Teams",
+      "View Team",
+      "Members",
+    ]),
+    ...BuildBreadcrumbLinksByTitles(PageMap.SETTINGS_TEAM_VIEW_PERMISSIONS, [
+      "Project",
+      "Settings",
+      "Teams",
+      "View Team",
+      "Permissions",
+    ]),
+    ...BuildBreadcrumbLinksByTitles(
+      PageMap.SETTINGS_TEAM_VIEW_BLOCK_PERMISSIONS,
+      ["Project", "Settings", "Teams", "View Team", "Block Permissions"],
+    ),
+    ...BuildBreadcrumbLinksByTitles(PageMap.SETTINGS_TEAM_VIEW_COMPLIANCE, [
+      "Project",
+      "Settings",
+      "Teams",
+      "View Team",
+      "Compliance",
+    ]),
+    ...BuildBreadcrumbLinksByTitles(PageMap.SETTINGS_TEAM_VIEW_DELETE, [
+      "Project",
+      "Settings",
+      "Teams",
+      "View Team",
+      "Delete",
+    ]),
     ...BuildBreadcrumbLinksByTitles(PageMap.SETTINGS_USERS, [
       "Project",
       "Settings",
@@ -44,6 +76,27 @@ export function getSettingsBreadcrumbs(path: string): Array<Link> | undefined {
       "Settings",
       "Users",
       "View User",
+    ]),
+    ...BuildBreadcrumbLinksByTitles(PageMap.SETTINGS_USER_VIEW_TEAMS, [
+      "Project",
+      "Settings",
+      "Users",
+      "View User",
+      "Teams",
+    ]),
+    ...BuildBreadcrumbLinksByTitles(PageMap.SETTINGS_USER_VIEW_CUSTOM_FIELDS, [
+      "Project",
+      "Settings",
+      "Users",
+      "View User",
+      "Custom Fields",
+    ]),
+    ...BuildBreadcrumbLinksByTitles(PageMap.SETTINGS_USER_VIEW_DELETE, [
+      "Project",
+      "Settings",
+      "Users",
+      "View User",
+      "Remove",
     ]),
     ...BuildBreadcrumbLinksByTitles(PageMap.SETTINGS_NOTIFICATION_SETTINGS, [
       "Project",
