+++ App/FeatureSet/Dashboard/src/Utils/PageMap.ts
@@ -511,8 +511,18 @@ enum PageMap {
   // Team
   SETTINGS_TEAMS = "SETTINGS_TEAMS",
   SETTINGS_USERS = "SETTINGS_USERS",
+
   SETTINGS_USER_VIEW = "SETTINGS_USER_VIEW",
+  SETTINGS_USER_VIEW_TEAMS = "SETTINGS_USER_VIEW_TEAMS",
+  SETTINGS_USER_VIEW_CUSTOM_FIELDS = "SETTINGS_USER_VIEW_CUSTOM_FIELDS",
+  SETTINGS_USER_VIEW_DELETE = "SETTINGS_USER_VIEW_DELETE",
+
   SETTINGS_TEAM_VIEW = "SETTINGS_TEAM_VIEW",
+  SETTINGS_TEAM_VIEW_MEMBERS = "SETTINGS_TEAM_VIEW_MEMBERS",
+  SETTINGS_TEAM_VIEW_PERMISSIONS = "SETTINGS_TEAM_VIEW_PERMISSIONS",
+  SETTINGS_TEAM_VIEW_BLOCK_PERMISSIONS = "SETTINGS_TEAM_VIEW_BLOCK_PERMISSIONS",
+  SETTINGS_TEAM_VIEW_COMPLIANCE = "SETTINGS_TEAM_VIEW_COMPLIANCE",
+  SETTINGS_TEAM_VIEW_DELETE = "SETTINGS_TEAM_VIEW_DELETE",
 
   // Workspace integrations
   SETTINGS_SLACK_INTEGRATION = "SETTINGS_SLACK_INTEGRATION",
